03_my_implementations/chapter4_serving_patterns/07_sync_async_pattern/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ML
onnxruntime==1.20.1
numpy==2.2.0
pillow==11.0.0

# Redis
redis==5.2.1

# Utilities
pydantic==2.10.3
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.3
mypy==1.13.0
